+++ tests/conftest.py
@@ -44,8 +44,10 @@ def app(settings: Settings):
 
 
 @pytest.fixture()
-def client(app):
-    with TestClient(app) as test_client:
+def client(app, settings: Settings):
+    with TestClient(
+        app, headers={"X-API-Key": settings.api_key}
+    ) as test_client:
         yield test_client